transfer

+++ src/main/java/org/example/mybank/entity/myObject/transaction_param.java
@@ -3,11 +3,13 @@ package org.example.mybank.entity.myObject;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class transaction_param {
     private String accountNumber;
     private String password;
-    private Integer amount;
+    private double amount;
     private String toAccountNumber;
     private String toAccountName;
 }
